+++ src/com/intellij/idea/plugin/hybris/common/utils/HybrisIcons.kt
@@ -1,5 +1,5 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
@@ -111,6 +111,7 @@ object HybrisIcons {
     val TS_GROUP_ITEM = getIcon("/icons/typeSystem/groupByItem.svg")
     val TS_GROUP_MAP = getIcon("/icons/typeSystem/groupByMap.svg")
     val TS_GROUP_RELATION = getIcon("/icons/typeSystem/groupByRelation.svg")
+    val TS_DESCRIPTION = AllIcons.Windows.Help
     val TS_ENUM = getIcon("/icons/typeSystem/enum.svg")
     val TS_ENUM_VALUE = getIcon("/icons/typeSystem/enumValue.svg")
     val TS_ATOMIC = getIcon("/icons/typeSystem/atomic.svg")
@@ -130,6 +131,13 @@ object HybrisIcons {
     val TS_IMPORT = AllIcons.ToolbarDecorator.Import
     val TS_ANNOTATION = AllIcons.Nodes.Annotationtype
     val TS_HEADER_ABBREVIATION = getIcon("/icons/typeSystem/headerAbbreviation.svg")
+    val TS_ALTERNATIVE_DECLARATION = AllIcons.Actions.Forward
+    val TS_SIBLING = AllIcons.Gutter.OverridenMethod
+    val TS_PREVIEW_SHOW = AllIcons.Actions.Show
+    val TS_PREVIEW_SHOW_COLLECTIONS = AllIcons.Actions.GroupByPrefix
+    val TS_PREVIEW_SHOW_ENUMS = AllIcons.Actions.GroupByTestProduction
+    val TS_PREVIEW_SHOW_MAPS = AllIcons.Actions.GroupByPackage
+    val TS_PREVIEW_SHOW_ITEMS = AllIcons.Actions.GroupByClass
 
     val BS_GROUP_BY_BEAN_DTO = getIcon("/icons/beanSystem/groupByDTO.svg")
     val BS_GROUP_BY_BEAN_EVENT = getIcon("/icons/beanSystem/groupByEvent.svg")
@@ -143,6 +151,14 @@ object HybrisIcons {
     val BS_ENUM = getIcon("/icons/beanSystem/enum.svg")
     val BS_ENUM_VALUE = getIcon("/icons/beanSystem/enumValue.svg")
     val BS_LEVEL_MAPPING = getIcon("/icons/beanSystem/levelMapping.svg")
+    val BS_ALTERNATIVE_DECLARATION = AllIcons.Actions.Forward
+    val BS_SIBLING = AllIcons.Gutter.OverridenMethod
+    val BS_PREVIEW_SHOW = AllIcons.Actions.Show
+
+    val CODE_NOT_GENERATED = AllIcons.General.ExclMark
+
+    val DIAGRAM = AllIcons.FileTypes.Diagram
+    val DIAGRAM_DIFF = AllIcons.Actions.DiagramDiff
 
     val BP_DIAGRAM_WAIT = getIcon("/icons/businessProcess/diagram/wait.svg")
     val BP_DIAGRAM_END = getIcon("/icons/businessProcess/diagram/end.svg")
@@ -159,6 +175,7 @@ object HybrisIcons {
     val BP_DIAGRAM_CLASS = AllIcons.Nodes.Class
     val BP_DIAGRAM_PARAMETER_REQUIRED = AllIcons.Nodes.Plugin
     val BP_DIAGRAM_PARAMETER_OPTIONAL = AllIcons.Nodes.Pluginobsolete
+    val BP_DIAGRAM_PROPERTIES = AllIcons.Nodes.Property
 
     val TS_DIAGRAM_PROPERTY = AllIcons.Nodes.Property
     val TS_DIAGRAM_DEPLOYMENT = AllIcons.Debugger.Db_db_object
@@ -181,11 +198,15 @@ object HybrisIcons {
 
     val GUTTER_POPULATOR = getIcon("/icons/gutter/populator.svg")
 
+    val CONSOLE = AllIcons.Debugger.Console
     val CONSOLE_SOLR = getIcon("/icons/console/solr.svg")
     val CONSOLE_OPEN = getIcon("/icons/console/open.svg")
     val CONSOLE_EXECUTE = AllIcons.Actions.Execute
     val CONSOLE_SUSPEND = AllIcons.Actions.Suspend
     val CONSOLE_EXECUTE_COMMIT_MODE_OFF = getIcon("/icons/console/executeWithCommitModeOff.svg")
+
+    val CONNECTION_ADD = AllIcons.General.Add
+    val CONNECTION_EDIT = AllIcons.Actions.Edit
 
     val TABLE_COLUMN_INSERT_LEFT = getIcon("/icons/table/columnInsertLeft.svg")
     val TABLE_COLUMN_INSERT_RIGHT = getIcon("/icons/table/columnInsertRight.svg")
@@ -200,6 +221,17 @@ object HybrisIcons {
     val UPLOAD = getIcon("/icons/upload.svg")
     val DELETE = getIcon("/icons/delete.svg")
     val CANCEL = DELETE
+
+    val ACTION_COLLAPSE_ALL = AllIcons.Actions.Collapseall
+    val ACTION_EXPAND_ALL = AllIcons.Actions.Expandall
+    val ACTION_CLEAR_ALL = AllIcons.Actions.GC
+    val ACTION_REMOVE = AllIcons.General.Remove
+    val ACTION_FORCE_REFRESH = AllIcons.Actions.ForceRefresh
+    val ACTION_FORWARD = AllIcons.Actions.Forward
+
+    val SCOPE_LOCAL = AllIcons.Ide.LocalScope
+
+    val NODE_JUNK = AllIcons.Modules.ExcludedGeneratedRoot
 
     private fun getIcon(path: String) = ReflectionUtil.getGrandCallerClass()
         ?.let { IconLoader.getIcon(path, it) }