+++ src/com/intellij/idea/plugin/hybris/diagram/businessProcess/BpDiagramNodeContentManager.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -21,8 +21,8 @@ package com.intellij.idea.plugin.hybris.diagram.businessProcess
 import com.intellij.diagram.AbstractDiagramNodeContentManager
 import com.intellij.diagram.DiagramBuilder
 import com.intellij.diagram.DiagramCategory
-import com.intellij.icons.AllIcons
 import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils.message
+import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.diagram.businessProcess.node.graph.BpGraphFieldContextParameter
 import com.intellij.idea.plugin.hybris.diagram.businessProcess.node.graph.BpGraphFieldParameter
 
@@ -37,7 +37,7 @@ class BpDiagramNodeContentManager : AbstractDiagramNodeContentManager() {
     }
 
     companion object {
-        val PARAMETERS = DiagramCategory({ message("hybris.diagram.bp.provider.category.parameters") }, AllIcons.Nodes.Property, true, false)
+        val PARAMETERS = DiagramCategory({ message("hybris.diagram.bp.provider.category.parameters") }, HybrisIcons.BP_DIAGRAM_PROPERTIES, true, false)
         val CATEGORIES = arrayOf(PARAMETERS)
     }