+++ src/com/intellij/idea/plugin/hybris/diagram/module/execution/lineMarker/ShowModuleDepDiagramRunLineMarkerContributor.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -18,8 +18,8 @@
 package com.intellij.idea.plugin.hybris.diagram.module.execution.lineMarker
 
 import com.intellij.execution.lineMarker.RunLineMarkerContributor
-import com.intellij.icons.AllIcons
 import com.intellij.idea.plugin.hybris.common.HybrisConstants
+import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.system.extensioninfo.model.ExtensionInfo
 import com.intellij.openapi.actionSystem.ActionManager
 import com.intellij.psi.PsiElement
@@ -42,6 +42,6 @@ class ShowModuleDepDiagramRunLineMarkerContributor : RunLineMarkerContributor() 
         if (DomManager.getDomManager(xmlFile.project).getFileElement(xmlFile, ExtensionInfo::class.java) == null) return null
 
         val action = ActionManager.getInstance().getAction("ShowModuleDependencyDiagram") ?: return null
-        return Info(AllIcons.FileTypes.Diagram, arrayOf(action)) { action.templateText }
+        return Info(HybrisIcons.DIAGRAM, arrayOf(action)) { action.templateText }
     }
 }