+++ src/com/intellij/idea/plugin/hybris/project/configurators/impl/DefaultSearchScopeConfigurator.java
@@ -1,7 +1,7 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -20,7 +20,6 @@
 package com.intellij.idea.plugin.hybris.project.configurators.impl;
 
 import com.intellij.find.FindSettings;
-import com.intellij.icons.AllIcons;
 import com.intellij.ide.projectView.impl.ModuleGroup;
 import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils;
 import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons;
@@ -94,7 +93,7 @@ public class DefaultSearchScopeConfigurator implements SearchScopeConfigurator {
             newScopes.add(hybrisScope);
         }
         if (groupExists(model, nonHybrisGroupName)) {
-            newScopes.add(createScope(AllIcons.Ide.LocalScope, nonHybrisGroupName));
+            newScopes.add(createScope(HybrisIcons.INSTANCE.getSCOPE_LOCAL(), nonHybrisGroupName));
         }
         newScopes.add(new NamedScope(
             HybrisI18NBundleUtils.message("hybris.scope.editable.all.ts.files"),